feat: finish main activity

+++ app/src/main/java/com/example/weather/model/CityModel.kt
@@ -6,5 +6,5 @@ data class CityModel(
     val picPath: String,
     val wind: Int,
     val humidity: Int,
-    val rain: Int
+    val rain: Double
 )
